+++ spec/acceptance/configuration_modification_spec.rb
@@ -48,6 +48,13 @@ module Transpec
         File.read('spec/spec_helper.rb').should == <<-END
           RSpec.configure do |config|
             config.mock_with :rspec do |mocks|
+              # In RSpec 3, `any_instance` implementation blocks will be yielded the receiving
+              # instance as the first block argument to allow the implementation block to use
+              # the state of the receiver.
+              # In RSpec 2.99, to maintain compatibility with RSpec 3 you need to either set
+              # this config option to `false` OR set this to `true` and update your
+              # `any_instance` implementation blocks to account for the first block argument
+              # being the receiving instance.
               mocks.yield_receiver_to_any_instance_implementation_blocks = true
             end
           end
@@ -97,6 +104,13 @@ module Transpec
             File.read(spec_helper_path).should == <<-END
           RSpec.configure do |config|
             config.mock_with :rspec do |mocks|
+              # In RSpec 3, `any_instance` implementation blocks will be yielded the receiving
+              # instance as the first block argument to allow the implementation block to use
+              # the state of the receiver.
+              # In RSpec 2.99, to maintain compatibility with RSpec 3 you need to either set
+              # this config option to `false` OR set this to `true` and update your
+              # `any_instance` implementation blocks to account for the first block argument
+              # being the receiving instance.
               mocks.yield_receiver_to_any_instance_implementation_blocks = true
             end
           end